# specs/001-sistema-gestion-consultorio/contracts/api.yaml
openapi: 3.1.0
info:
  title: Salud desde el Alma API
  version: 0.1.0
  description: >-
    Contrato inicial para panel administrativo, agenda, pagos y webhook de WhatsApp
    del MVP de Salud desde el Alma.
servers:
  - url: https://api.saluddesdeelalma.org
    description: Produccion
  - url: http://localhost:4000
    description: Desarrollo local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sda_admin_session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
...
    Appointment:
      type: object
      required:
        [
          id,
          scheduledAt,
          endsAt,
          therapyType,
          durationMinutes,
          modality,
          status,
          isManualException,
          locationLabel,
          meetingLink,
          cancelReason,
          cancelledAt,
          cancellationNotice,
          createdVia,
          paymentStatus,
          payments,
          patientId,
          patientName,
          patientPhone,
          patientBirthdate,
          therapistId,
          therapistName,
          therapistIsActive
        ]
      properties:
        id:
          type: string
          format: uuid
        scheduledAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        therapyType:
          type: string
          enum: [individual, pareja, familiar]
        durationMinutes:
          type: integer
          enum: [60, 90]
        modality:
          type: string
          enum: [online, presencial]
        status:
          type: string
          enum: [programada, confirmada, completada, cancelada]
        createdVia:
          type: string
          enum: [whatsapp, panel, system]
        isManualException:
          type: boolean
        locationLabel:
          type: string
          nullable: true
        meetingLink:
          type: string
          format: uri
          nullable: true
        cancelReason:
          type: string
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancellationNotice:
          type: string
          enum: [a_tiempo, tardia]
          nullable: true
        paymentStatus:
          type: string
          enum: [pendiente, anticipo, completado]
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        patientId:
          type: string
          format: uuid
        patientName:
          type: string
        patientPhone:
          type: string
        patientBirthdate:
          type: string
          format: date
          nullable: true
        therapistId:
          type: string
          format: uuid
          nullable: true
        therapistName:
          type: string
          nullable: true
        therapistIsActive:
          type: boolean
    Payment:
      type: object
      required:
        [
          id,
          paymentType,
          amount,
          method,
          status,
          proofReference,
          paidAt,
          createdAt
        ]
      properties:
        id:
          type: string
          format: uuid
        paymentType:
          type: string
          enum: [anticipo, completo]
        amount:
          type: number
          format: float
        method:
          type: string
          enum: [transferencia, efectivo]
        status:
          type: string
          enum: [pendiente_validacion, validado, rechazado]
        proofReference:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/appointments:
    get:
      tags: [Appointments]
      summary: Lista citas en un rango ISO (solo cuenta `admin`)
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Citas del calendario en el rango solicitado
          content:
            application/json:
              schema:
                type: object
                required: [appointments]
                properties:
                  appointments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Appointment"
    post:
      tags: [Appointments]
      summary: Crea una cita desde el panel
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduledAt, modality, therapyType]
              anyOf:
                - required: [patientId]
                - required: [patient]
              properties:
                patientId:
                  type: string
                  format: uuid
                patient:
                  type: object
                  required: [fullName, whatsappPhone, birthdate]
                  properties:
                    fullName:
                      type: string
                    whatsappPhone:
                      type: string
                    birthdate:
                      type: string
                      format: date
                    preferredModality:
                      type: string
                      enum: [online, presencial]
                    email:
                      type: string
                      format: email
                scheduledAt:
                  type: string
                  format: date-time
                modality:
                  type: string
                  enum: [online, presencial]
                therapyType:
                  type: string
                  enum: [individual, pareja, familiar]
                manualExceptionConfirmed:
                  type: boolean
                  default: false
                isManualException:
                  type: boolean
                  default: false
                locationLabel:
                  type: string
                  maxLength: 255
                meetingLink:
                  type: string
                  format: uri
      responses:
        "201":
          description: Cita creada
          content:
            application/json:
              schema:
                type: object
                required: [appointment]
                properties:
                  appointment:
                    $ref: "#/components/schemas/Appointment"
        "409":
          description: El intervalo solicitado se traslapa con una cita activa de la psicóloga asignada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/appointments/{appointmentId}:
    patch:
      tags: [Appointments]
      summary: Reagenda una cita
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduledAt]
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                modality:
                  type: string
                  enum: [online, presencial]
                therapyType:
                  type: string
                  enum: [individual, pareja, familiar]
      responses:
        "200":
          description: Cita reagendada
          content:
            application/json:
              schema:
                type: object
                required: [appointment]
                properties:
                  appointment:
                    $ref: "#/components/schemas/Appointment"
  /api/v1/appointments/{appointmentId}/confirm:
    post:
      tags: [Appointments]
      summary: Confirma una cita programada y despacha su confirmacion
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cita confirmada
          content:
            application/json:
              schema:
                type: object
                required: [appointment]
                properties:
                  appointment:
                    $ref: "#/components/schemas/Appointment"
        "409":
          description: La cita no esta programada o no puede confirmarse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/appointments/{appointmentId}/cancel:
    post:
      tags: [Appointments]
      summary: Cancela una cita y programa la notificacion al paciente
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Cita cancelada
          content:
            application/json:
              schema:
                type: object
                required: [appointment]
                properties:
                  appointment:
                    $ref: "#/components/schemas/Appointment"
...